requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
deep-translator==1.4.4
easyocr==1.3.2
google-trans-new==1.1.9
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
langdetect==1.0.9
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.2.52
Pillow==8.2.0
pyparsing==2.4.7
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
urllib3==1.26.5
